Moved logic from plugin to core
