+++ patcher/.gitignore
@@ -1,6 +1,8 @@
-.vs/*
-bin/*
-obj/*
+.vs
+bin
+obj
 /HitmanPatcher.sln.DotSettings.user
 /.idea/.idea.HitmanPatcher/.idea/riderMarkupCache.xml
 *.suo
+*.user
+/Publish
