+++ static/peacock-menu/index.json
@@ -6,7 +6,7 @@
         },
         "do": {
             "$include": {
-                "$path": "$formatstring menusystem/plugins/peacock-menu/options.json@{$.@global.CacheBuster}"
+                "$path": "$formatstring menusystem/pages/peacock-menu/options.json@{$.@global.CacheBuster}"
             }
         }
     }
