+++ static/peacock-menu/options.json
@@ -46,7 +46,7 @@
         },
         {
             "$include": {
-                "$path": "$formatstring menusystem/plugins/peacock-menu/flags/index.json@{$.@global.CacheBuster}"
+                "$path": "$formatstring menusystem/pages/peacock-menu/flags/index.json@{$.@global.CacheBuster}"
             }
         }
     ]
